src/ac_training_lab/xarm5/requirements.txt
# XArm5 Raspberry Pi Control Requirements

# Official UFACTORY xArm Python SDK
xarm-python-sdk>=1.17.0

# Core dependencies (usually included with SDK)
numpy>=1.20.0
